Reimplement copy-webpack-plugin for vite
We have a few dependencies for which we copy static assets to
`public/assets/webpack`: WebIDE, Source Graph and the PDF viewer.

In webpack we use the `copy-webpack-plugin` for this. It basically takes
an array of files/directories and copies them to
`public/assets/webpack`.

In this commit we move the definition for this to `webpack.constants.js`
and import it to `vite.config.js`. Given our config for the
`copy-webpack-plugin` is rather limited (only absolute paths, no globs),
we can re-implement a rather smallish scope of the plugin for vite.