Add `deprecate_vulnerability_occurrence_pipelines` feature flag
Context
------------

We want to drop the `vulnerability_occurrence_pipelines` table. We
have some features that currently use this table. We are migrating
those features to use new columns before dropping the table. This is
all tracked in [this epic][0]

We will need to run the existing implementation and the new
implementation side-by-side for a bit to give ourselves a safe
roll-back path with no data loss.

This commit
------------

1. Adds a new [`wip` feature flag][1] called
   `deprecate_vulnerability_occurrence_pipelines`

2. Updates the `vulnerabilities/findings` factory to support both
   implementations side-by-side

3. Begins using thed FF in the findings model to toggle between the
   two implementations

This feature flag will be transitioned to a `gitlab_com_derisk`
feature flag once all of the application code changes are complete and
the data backfills are complete

[0]:https://gitlab.com/groups/gitlab-org/-/epics/11241
[1]:https://docs.gitlab.com/ee/development/feature_flags/index.html#wip-type

related to: https://gitlab.com/gitlab-org/gitlab/-/issues/422384
resolves: https://gitlab.com/gitlab-org/gitlab/-/work_items/443495
resolves: https://gitlab.com/gitlab-org/gitlab/-/work_items/450945