Strip namespaces from OS component names
The namespace is included for most OS package types, and if not removed,
will cause findings created by Continous Container Scanning to deviate
from the output format established by the container-scanning analyzer.
To maintain parity, we now remove the namespace from specific PURL
types. Namely, we remove the namespace from components with the `apk`,
`deb`, and `rpm` PURL types. This means that we'll now parse a component
name like `debian/curl` and transform it to `curl` which is the correct
behavior.

Changelog: fixed
EE: true