Add test validating file line endings not mutated
When a repository file is created or modified in GitLab it is expected
that the file content is committed exactly as submitted. Previously,
Gitaly used the Git `core.autocrlf=input` configuration which mutated
file line endings to always be `LF` when Gitaly generates the commit.
This configuration has now been removed from Gitaly. Therefore, file
line endings are no longer mutated.

Add a test validating that a file with `CRLF` line endings is committed
without being modified.