Hide invited group name and source from project/group non-admins
If a group invited to a project/group is not visible to the current user
we mask the source. The visibility was determined by:

1. The current user can read the invited group.
2. The current user is a member of the shared group.

We're changing point 2 above to:

The current user is the admin of the shared group/project i.e. having at
least maintainer access in the shared project or having owner access in
the shared group.

Changelog: changed