Fix incorrect deprecation item title
We recently deprecated the `omnibus_gitconfig` configuration key in
GitLab 16.10. The initial version of the change was referring to the
key by using the internal name in Omnibus rather than the key the end
user would be configuring. The body of the deprecation item was updated
but the old key was left in the title. Fix the title to refer to the
correct key as well.