Fix formatting of issue summaries
In some summaries, we found random numbers within tags or closing the
summary with """. This was due to the format we requested from the LLM.

With the new prompt we separate each comment within a <comment> tag,
remove the random number and the """ at the end of the prompt.

Changelog: changed
EE: true