Use semaphore for storing security reports
If many StoreSecurityReportsWorker jobs are running at the same time,
this can cause database queries to timeout while they are waiting for
locks to be released. This MR performs locking at the application level
on a per-project basis, so that we don't experience these locks during
database transactions.

Changelog: fixed
EE: true