Update `clickup.md` with custom task Ids info
Following *Live Chat Follow-up: with a clickup representative*, this is
their recommendation

```

This is because Custom Task IDs require your Workspace/team ID to be
present as well.

After some further testing, I can confirm that if you use the format
below, it will then generate links for both Custom Task IDs and normal
task IDs:

https://app.clickup.com/t/WorkspaceID/:id

You need to replace the WorkspaceID part with your Workspace ID which
you can get by simply logging in to your ClickUp account it's the
numbers right after app.clickup.com, so as an example for mine:

```

Therefore, although the  `t/:id` format still works, since the
`t/workspaceId/:id` is more generic, I think it makes sense to just
recommend that second one, instead of explaining the differences and
when to use one vs the other.