Fix handling of DAGs within a stage in AtomicProcessingService
Suppose you had a `.gitlab-ci.yml` with a manual job listed after a
`needs`:

```
test1:
  needs: [manual1]
  script: exit 0

manual1:
  stage: test
  when: manual
  script: exit 0
```

Previously the `test1` job would be stuck in the `created` state
instead of `skipped` because `AtomicProcessingService` did not
consider a DAG within a stage. `test1` only gets updated to `skipped`
if `Ci::ProcessBuildService` is called on `manual1` first, but since
`test1` is listed first `AtomicProcessingService` will usually call
`Ci::ProcessBuildService` on it first. The update order previously
depended on the order returned by the database in
`StatusCollection#all_jobs`.

This commit introduces a `ci_atomic_processing_ordered_update_stage`
feature flag that performs a topological sort on all jobs in the stage
and updates the jobs in that order. This is similar to the approach
taken in `ResetSkippedJobsService`.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/450395

Changelog: fixed