Fix flaky spec for epic work item validation
It can happen that the work_item and the epic have different
`updated_at` values. In EpicWorkItemSync::Diff we account for this and
allow a certain time difference.

However, when the threshold exceeds, it can happen that `updated_at` is
part of the the mismatched_attributes.

We cover this as part of the EpicWorkItemSync::Diff tests already. To
not have a flaky spec again in the Worker, we only ensure that the title
and lock_version attributes are part of the mismatched_attributes
instead of a strict match of the array.