Fix start_is_fixed and due_date_is_fixed default
When creating a work item where the `start_date` or `due_date` param was
set but `nil`, we still set `start_date_is_fixed` and
`due_date_is_fixed` as `true`, even when the parameter was set to
`false`.

This caused an issue when syncing epic work items, where we created
`dates_source` with the wrong default values.

Changelog: fixed
EE: true