ActionController::UrlGenerationError for old path redirects
Contributes to https://gitlab.com/gitlab-org/gitlab/-/issues/452061

**Problem**

When project path was changed, then we try to redirect all requests from
the old path to the new path.

But it's possible that some parameters of the request are incorrect and
the generation of new path URL fails.

As a result, we return a 500 error to the user.

**Solution**

Add an error handler for this case. If the URL generation is not
successful, then display a 404 page.

Changelog: fixed