Quarantine flaky atomic processing ResetSkippedJobsService specs
This quarantines two flaky specs that failed since they depend on the
order of CI jobs returned from the database.

The fix for this bug is implemented in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/147875, but this
will not be backported, since it's not a crucial fix and introduces
a feature flag.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/450395