Handle controller paths with non-existant entry points
We have the controller `profiles/keys` with the action `show`, so it
tries to load: `pages.profiles.keys.show.js` which doesn't exist. The
following entry points do exist:

   pages/profiles/keys/show/index.js <== does not exist
   pages/profiles/keys/index.js <== does exist
   pages/profiles/index.js <== does exist

So, it should have tried loading the other two, but didn't. By moving
the "backtracking" logic to the backend, we can try loading all of those
virtual entry points.