Rename `NamespaceSettings::UpdateService` class
This commit refactors the NamespaceSettings::UpdateService class
to NamespaceSettings::AssignAttributesService, since it does not
actually updated namespace settings, it assigns attributes.

The commit involves a complete refactoring of the class and all
references to it in other files, along with the renaming of any
ruby namespace file with "update_service" in the name to "assign_
attributes_service", and updating rubocop config files that
reference them. This is done in both the core and enterprise
directories.

Fixes https://gitlab.com/gitlab-org/gitlab/-/issues/444223