Track component usage in create_pipeline_service
Adds a Ci::Pipeline::Chain step 'TemplateUsage' that
tracks the usage of CI Catalog components in the pipeline.
It first finds if any of the included components exist in the
CI Catalog. Then it sends an InternalEvent for each catalog
component found. This change is made behind
the feature flag `ci_track_component_usage_internal_event`.