Update timebox report related specs
The data for charting a burndown/up chart are a timeseries.

In the real word, the chart data are generated based on
the events generated in the past.

For example, "an issue was assigned to a milestone"
and "an issue was closed", etc.

Some specs used fake "future" events to generate and test timeseries.
This commit updates them to generate and test a past timeseries.