Add empty state for sidebar CRM widget
The sidebar CRM widget for issues currently has no empty state, which is
not aligned with the experience for other sidebar widgets.

This MR adds an empty state similar to the temporary time tracking
sidebar widget for work items, which prompts the user to use a quick
action to manage that widget's contents. A future MR will explore GUI
controls for CRM contacts on issues.

Changelog: changed