Fix duplicate "Job suceeded lines"
`Projects::JobsController#trace` reads the last 512 kilobytes of a CI
job trace and attempts to parse it. Previously if the last part of
this trace omitted the `begin_section` marker, the `end_section`
marker detected that the section was not open and skipped it, without
incrementing the offset of the state. Subsequent lines would then have
the incorrect offset, causing the returned `state` to be short the
characters that were skipped. As a result, requests to
`Projects::JobsController#trace` would request an offset that preceded
the last line, causing duplicate lines to appear.

To fix this, if we detect an `end_section` marker that does not have a
matching `begin_section,` we just skip over it and advance the offset.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/452097

Changelog: fixed