Rename actual_head_pipeline to diff_head_pipeline
The "actual" in these method names was a misnomer. These methods all
retrieved the pipeline by `MergeRequest#diff_head_sha`, which is not
updated at the same time as the repository. This is an important
distinction, because branch pipeline can be created for a merge
request's source branch _before_ the merge request diff has been
updated.