Fix URL validator for mirror services when using localhost
After converting from the old UrlBlocker to the new one
(https://gitlab.com/gitlab-org/gitlab/-/issues/430830),
we saw that some self-hosted users experienced problems with the mirror
repositories (https://gitlab.com/gitlab-org/gitlab/-/issues/443296).

The problem is caused when using local domains. I investigated the
problem and found that UpdateMirrorService and UpdateRemoteMirrorService
do not send the allow_localhost and allow_local_network settings
to UrlBlocker. After adding these settings, it started to work
successfully.

It's still strange that this used to be working in the old UrlBlocker.

Changelog: fixed