Revise page for flow and fix line-level problems
This page needed some CTRT love, and some fixes to its ordered lists.
The lists are now complete, and I've blended a non-task heading
back into the concept area. I've also reworked the flow of the
various subheadings somewhat, and done a lot of line-level polish
to reduce the amount of passive voice.