Add new page for create custom roles
Currently on the Roles and Permissions page, 
when the Add new role button is clicked,
the new role form is shown directly on the page. 
Because the form is complex, 
to match the behavior of other complex forms in the app, 
the new role form has been moved to its own page.