Add API endpoint to return current SSO session expiry time
This change will add an API endpoint on EE to return the expiry time of
the user's current Group SAML SSO sessions. This way the frontend can be
more responsive if a tab is left open overnight or over the weekend, and
the client no longer has an active SSO session.

Changelog: added
EE: true