Fix merge request suggestions adding incorrect `+`
Fixes a bug where if you add a suggestion on the changes tab and then
switch to the overview tab, click reply and add a new suggestion it
will add an invalid `+` symbol at the start.
This happens because we don't call `prepareDiffLines` when a new note is added,
only when notes are fetched.

https://gitlab.com/gitlab-org/gitlab/-/issues/454227