Fix non-unique :key with multiple project keys
When we only had a single project key, we extracted an `id` from
the relative reference. For example, FE-123 would have `id` as 123.

This can cause issues with multiple project keys as issues can
have duplicate `id`, causing them to have the same `:key` binding
and not render properly. For example, BE-123 and FE-123 would
have the same `id` and Vue would re-use the component (and
data) for both.

This fix changes the id to use the Jira reference directly to
avoid such issues.