Migrate inputs to spec on CI components table
`catalog_resource_components` was created with an `inputs` column that
stores the inputs value from a CI template's header. However, the
components table now needs to store CI steps as well as templates, and
steps have outputs as well as inputs. To handle this, a new `spec`
column was introduced that stores the entire header.

This commit migrates the values from the `inputs` column of existing CI
component records to the `spec` column so that we can remove the
`inputs` column in the next milestone. It appends an `inputs:` key to
the existing inputs hash so it has the correct form for the `spec`
column.

Changelog: other
Issue: https://gitlab.com/gitlab-org/gitlab/-/issues/443662