Fix errors in CI JSON Schema
Spotted using 'cue import' to translate the JSON Schema to CUE
(https://cuelang.org). Previously reported via
https://cuelang.org/issue/1939.

'minLength' is a constraint that can only be used on strings:

    https://json-schema.org/understanding-json-schema/reference/string#length

Arrays should be constrained with 'minItems':

    https://json-schema.org/understanding-json-schema/reference/array#length

The removal of the stray 'additionalProperties' on the
'patternProperties' field object is also correct, on the basis an
'patternProperties' object is defined as being a map from field name
pattern to type.

Signed-off-by: Paul Jolly <paul@myitcv.io>