Show banner when there is a seat overage
When block seat overages is enabled, and there are more members in the
namespace than there are seats in the subscription, display a
banner alert informing that the namespace is now read-only.
Inform owners that they can unlock the namespace by
reducing the number of members or buying more seats.
Inform non-owners to contact their group owner.
Block seat overages is behind a feature flag.