Refactor SBoM ingestion hash lookup
This change removes duplicated hash lookups in each SBoM ingestion task
by adding an `each_pair` method. This method iterates through the
`return_data` of a task and yields the row with its corresponding
`occurrence_map`. It does this by creating a hash index based on the
`unique_by` attributes.