Remove usages of FullPath and SetFullPath RPCs
Contributes to https://gitlab.com/gitlab-org/gitlab/-/issues/408940

**Problem**

Gitaly plans to remove the SetFullPath and FullPath RPCs in the context
of https://gitlab.com/groups/gitlab-org/-/epics/8953. We need to adapt
Rails to stop calling those RPCs so that we can move ahead with the
planned removal.

**Solution**

Remove usages of FullPath and SetFullPath RPCs

Changelog: other