Skip committer check for commits signed by GitLab
When Web/UI commits signing is enabled, the committer email
of the commits created on GitLab instance have <noreply@gitlab.com>
committer emails.

Currently, the push rules either ignore the web
commits if possible or use author email instead because they
are the same for Web commits. This commit also adds an additional
condition: the commit must be signed by GitLab signature.

All the changes are behind a feature flag.