React to `archived` and `traversal_ids` changes while creating records
It's possible that the `project#archived` and `namespace#traversal_ids`
can change while we are creating the vulnerability records. If that
happens, the recently created `vulnerability_reads` records will have
incorrect `archived` and `traversal_ids` values.

To prevent aforementioned scenario, we check the project and namespace
objects after creating the records and schedule adjustment jobs if
there is a change.