Preserve indirect memberships
We import "shared" and "inherited shared" memberships from the source
https://gitlab.com/gitlab-org/gitlab/-/blob/39cae3e7c50bd508c9be35444b1f20159c051cbd/lib/bulk_imports/common/graphql/get_members_query.rb#L83-93
but previously we would create direct memberships on the destination
even if the user had a shared and inherited shared membership within
the destination.

This change preserves existing shared and inherited shared memberships
on the destination, unless the existing membership is of a lower
permission level than the source.

Changelog: fixed