[Batched Migration] Fix corrupted scanner_ids of vulnerability_reads
There was a bug in production where the `scanner_id` of a
vulnerability finding would be updated, yet the `scanner_id` of the
related `vulnerability_read` would remain unchanged. This was an issue
as the `vulnerability_read` should always be in sync with the
`vulnerability_finding`.

The bug has already been [fixed in production][0], so now we need a
migration to fix all of the `vulnerability_reads` that were affected
by the bug.

[0]:https://gitlab.com/gitlab-org/gitlab/-/merge_requests/148016

---

MR: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/148807
Related to: https://gitlab.com/gitlab-org/gitlab/-/issues/454290
resolves: https://gitlab.com/gitlab-org/gitlab/-/issues/454827

Changelog: fixed
EE: true