Document Cells rings
What
---
Copy most of the information around rings and cells from
[deployments.md](https://gitlab.com/gitlab-org/gitlab/-/blob/eec5c731a29bfdefd31b8b486900873d6a50340b/doc/architecture/blueprints/cells/infrastructure/deployments.md) to infrastructure index page.

Why
---
The ring definition proved to be useful for a lot of things not just
deployments and it's serving as a mental model of how we think about
Cells in production.

Adding information about this in the index page for infrastructure will
help reduce defining it multiple times.

Signed-off-by: Steve Xuereb <sxuereb@gitlab.com>