Add Tailwind base without preflight
This ensures that the Tailwind CSS bundle contains all the default CSS
variables declarations. This should pevent issues on Chrome where
`transform` style properties would be ignored due to only a subset of
variables being declared in the utils scope.

We also disable the `preflight` core plugin to ensure the base layer
doesn't come with all the resets we don't need/want.