Temporarily restore old breadcrumbs UX
The breadcrumbs are undergoing updates to improve the UX when the items
would overflow the viewport. These have caused a regression in existing
instances where the new behavior is not yet in place due to them not
using the Vue component. Until we are ready to implement the new
behavior there, this restores the previous one where the crumbs are
allowed to wrap on larger viewports, and are truncated on smaller ones.

Changelog: fixed