Quarantine a flaky test
The `Group member autocomplete when editing description of a group milestone for a member of a private group invited to the group suggests member of private group as well` test has been reported as flaky more then 1000 times.

This MR quarantines the test. This is a discussion starting point to let the responsible group know about the flakiness
so that they can take action:

- accept the merge request and schedule to improve the test
- close the merge request in favor of another merge request to delete the test

Related to https://gitlab.com/gitlab-org/gitlab/-/issues/444683.

This change was generated by
[gitlab-housekeeper](https://gitlab.com/gitlab-org/gitlab/-/tree/master/gems/gitlab-housekeeper)
using the Keeps::QuarantineFlakyTests keep.

To provide feedback on your experience with `gitlab-housekeeper` please comment in
<https://gitlab.com/gitlab-org/gitlab/-/issues/442003>.

Changelog: other