When localStorage draft exists for comment reply, show it on load
When a user has a comment draft saved in localstorage, we do not
indicate that on page load at all; the user can only see their comment
draft by clicking reply and seeing the draft populated into the markdown
editor.

This MR changes the noteable_discussion component's behavior such that
it exposes the markdown editor with the user's comment draft on page
load if there is a saved draft for the comment in localstorage.

Changelog: fixed