Purge CSS-in-JS config
This extends the tailwind_all_the_way.mjs to accept an --only-used
flag.

Also remove the `content` option override from the all_the_way config.

I don't think we need this anymore, we can rely on the conversion script
to decide whether or not we want to parse all the utils.