Use strong_params to permit service params
We are introducing more complex params for the branch rule services
which includes nested params. The current solution worked well for a
single level hash but fails to exclude params that are one or more nodes
deep.

We are utilizing ActionController::Parameters to help us wrangle these
paramters.

Part of https://gitlab.com/gitlab-org/gitlab/-/issues/441471