Obfuscate external participants emails in system notes
When managing external participants in an issue
we add system notes to display which email
was added or removed. On public issues or for
guest users this information should not be visible.

Adds obfuscation to issue email participant
system notes, so no email addresses are disclosed
if the user does not have the needed role.

Changelog: added