Utilize show_diff_preview_in_email column
Add the front-end updates needed to alter and utilize the
show_diff_preview_in_email column.

This should be used to hide the diff content found in the _note_email
templates.  Replace it with a string text informing the user that the
GitLab instance should be visited to see the line in question.

Changelog: added