Simplifies SemanticVersionable concern
Given that we should always be validating a version if we are going to
use this concern, we can remove the validation. We can also simplify
the getter and setter methods and remove the custom naming for
consistency.

This also simplifies the inclusion of the concern into the models.

Changelog: added