Show members of private group as approvers
When an approval rule has a group that is considered hidden to the
user then approvers are not shown.  This is a problem when a private
group has been added to an approval rule.

This change is to consider private groups that are invited to the
project associated with the approval rule to be considered non-hidden

Contributes to: https://gitlab.com/gitlab-org/gitlab/-/issues/281029
Changelog: fixed
EE: true