Error in rendering of token names
A bot user can have one pat at any point of time.
ExpiringWorker is written on the same premise and we expect
the user of type bot to have only one pat. This mailer need
not be an exception and run a loop on multiple token names

Changelog: fixed

Error in rendering of token names

Wrong code comitted