Bug fix for single line cancellations
When supporting running the after script on cancel we
need to add the job status to the runner update endpoint.
previously the status was only sent over on append and
the cancellation would cause to update endpoint to throw
a 403. Having only the append trace endpoint support sending
the job status causes single line scripts to never cancel
but works ok for scripts with multiple commands or with
multi-line output.