Deploy to release environment on all merges to stable branches
Change the rule to run the stage `release-environments` on all merges to
a stable branch, not only the ones that match `.setup-test-env-patterns`.

This way, a deployment to a release environment is also triggered when a
new release is created. The release creates a special commit that
changes the `VERSION` file, so it doesn't match
`.setup-test-env-patterns`.

Ref: https://gitlab.com/gitlab-com/gl-infra/delivery/-/issues/20163