Remove the `tailwind_all_the_way` feature flag
This removes the `tailwind_all_the_way` feature flag. We had the flag
enabled long enough in production to catch and fix a few regressions. We
are now ready to remove the flag and rely on a single Tailwind config
which will allow us to resume the migrations.

Changelog: added