Return or display Gitlab version if GITLAB_KAS_VERSION is a SHA
Gitlab::Kas.version and Gitlab::Kas.version_info are called
when the API or ~frontend needs to display the KAS semantic version.

Both Gitlab::Kas.version and Gitlab::Kas.version_info bases the version
on the contents of the GITLAB_KAS_VERSION file. We want to handle the
possibility that the content of the GITLAB_KAS_VERSION file is a SHA.
In this situation, Gitlab::Kas.version and Gitlab::Kas.version_info
will be based on the Gitlab version, with the SHA as the suffix.

Changelog: changed