Send tracking event for Duo Chat feedback from backend
Currently when a user submits feedback for a Duo Chat question, the
tracking event is sent from the front-end. This has the downside of
being caught by Ad/Tracking blockers on the client-side. We add the
ability for the backend to perform this task to avoid this. An upcoming
change will switch sending events from the frontend to the backend.