Quarantine a flaky test
The `TrialRegistrationsController POST create when on gitlab.com or in dev environment with snowplow tracking with email confirmation when email confirmation settings is not set to `soft` tracks an almost there redirect` test has the `flakiness::1` label set, which means it has more than 1000 flakiness reports.

This MR quarantines the test. This is a discussion starting point to let the responsible group know about the flakiness
so that they can take action:

- accept the merge request and schedule to improve the test
- close the merge request in favor of another merge request to delete the test

Related to https://gitlab.com/gitlab-org/gitlab/-/issues/436592.

This change was generated by
[gitlab-housekeeper](https://gitlab.com/gitlab-org/gitlab/-/tree/master/gems/gitlab-housekeeper)
using the Keeps::QuarantineFlakyTests keep.

To provide feedback on your experience with `gitlab-housekeeper` please comment in
<https://gitlab.com/gitlab-org/gitlab/-/issues/442003>.

Changelog: other