Add imported column to resources
This adds several migrations that add an imported column
to resources that will display an imported icon in the UI
if they are imported. The column is a smallint type that
will use an enum and default to 0. The default value 0
will indicate the resource was not imported. Any other
value will indicate the importer used (e.g. github,
bitbucket, etc.)

Changelog: added